update first version code
