+++ rdagent/scenarios/kaggle/experiment/meta_tpl/model_rf.py
@@ -9,19 +9,21 @@ import pandas as pd
 from sklearn.ensemble import RandomForestClassifier
 from sklearn.metrics import accuracy_score
 
-def select(X):
+
+def select(X: pd.DataFrame) -> pd.DataFrame:
     """
     Select relevant features. To be used in fit & predict function.
     """
     # For now, we assume all features are relevant. This can be expanded to feature selection logic.
     return X
 
+
 def fit(X_train: pd.DataFrame, y_train: pd.Series, X_valid: pd.DataFrame, y_valid: pd.Series):
     """
     Define and train the Random Forest model. Merge feature selection into the pipeline.
     """
     # Initialize the Random Forest model
-    model = RandomForestClassifier(n_estimators=100, random_state=32)  
+    model = RandomForestClassifier(n_estimators=100, random_state=32)
 
     # Select features (if any feature selection is needed)
     X_train_selected = select(X_train)
@@ -37,15 +39,16 @@ def fit(X_train: pd.DataFrame, y_train: pd.Series, X_valid: pd.DataFrame, y_vali
 
     return model
 
+
 def predict(model, X):
     """
     Keep feature selection's consistency and make predictions.
     """
     # Select features (if any feature selection is needed)
     X_selected = select(X)
-    
+
     # Predict using the trained model
     y_pred_prob = model.predict_proba(X_selected)[:, 1]
-    
+
     # Apply threshold to get boolean predictions
     return y_pred_prob > 0.5
